% --- Source file: ./man/evian_linear.Rd ---
  % Part of the evian package
% Distributed under GPL 2 or later

\name{evian_binary_bim}
\alias{evian_binary_bim}
\title{Example map data for evian_binary_raw.}
\description{
  This dataset stored the corresponding SNP information from \code{\link{evian_binary_raw}}.
}
\usage{
  data(evian_binary_bim)
}

\details{
  This is the corresponding map file for \code{\link{evian_binary_raw}}. Specifically it stores the chromosome, base pair, and two alleles for the 30 SNPs listed in \code{\link{evian_binary_raw}} in the same order.
}
