% --- Source file: ./man/evian_linear.Rd ---
  % Part of the evian package
% Distributed under GPL 2 or later

\name{evian_linear_bim}
\alias{evian_linear_bim}
\title{Example map data for evian_linear_raw.}
\description{
  This dataset stored the corresponding SNP information from \code{\link{evian_linear_raw}}.
}
\usage{
  data(evian_linear_bim)
}

\details{
  This is the corresponding map file for \code{\link{evian_linear_raw}}. Specifically it stores the chromosome, base pair, and two alleles for the 10 SNPs listed in \code{\link{evian_linear_raw}} in the same order.
}
