% --- Source file: ./man/evian_linear.Rd ---
  % Part of the evian package
% Distributed under GPL 2 or later

\name{evianmap_binary}
\alias{evianmap_binary}
\title{Example map data for eviandata_binary.}
\description{
  This dataset stored the corresponding SNP information from \code{\link{eviandata_binary}}.
}
\usage{
  data(evianmap_linear)
}

\details{
  This is the corresponding map file for \code{\link{eviandata_binary}}. Specifically it stores the chromosome, base pair, and two alleles for the 30 SNPs listed in \code{\link{eviandata_binary}} in the same order.
}
