% --- Source file: ./man/evian_linear.Rd ---
  % Part of the evian package
% Distributed under GPL 2 or later

\name{evianmap_linear}
\alias{evianmap_linear}
\title{Example map data for eviandata_linear.}
\description{
  This dataset stored the corresponding SNP information from \code{\link{eviandata_linear}}.
}
\usage{
  data(evianmap_linear)
}

\details{
  This is the corresponding map file for \code{\link{eviandata_linear}}. Specifically it stores the chromosome, base pair, and two alleles for the 10 SNPs listed in \code{\link{eviandata_linear}} in the same order.
}
